deployed_api/requirements.txt
flask
firebase-admin
functions_framework
google-cloud-firestore
firestore
python-dotenv
bcrypt
flask_cors
google-cloud
Werkzeug





